fix make lint
```
$ make lint
isort --check-only mahjong/*
black --check mahjong/*
All done! ✨ 🍰 ✨
92 files would be left unchanged.
flake8 mahjong --config .flake8
mahjong/hand_calculating/yaku.py:42:9: B028 No explicit stacklevel argument found. The warn method from the warnings module uses a stacklevel of 1 by default. This will only show a stack trace for the line on which the warn method is called. It is therefore recommended to use a stacklevel of 2 or greater to provide more information to the user.
mahjong/hand_calculating/yaku.py:47:9: B028 No explicit stacklevel argument found. The warn method from the warnings module uses a stacklevel of 1 by default. This will only show a stack trace for the line on which the warn method is called. It is therefore recommended to use a stacklevel of 2 or greater to provide more information to the user.
mahjong/meld.py:42:9: B028 No explicit stacklevel argument found. The warn method from the warnings module uses a stacklevel of 1 by default. This will only show a stack trace for the line on which the warn method is called. It is therefore recommended to use a stacklevel of 2 or greater to provide more information to the user.
make: *** [Makefile:8: lint] Error 1
```